gallium: MSVC warning fixes.
Conflicts:

	src/gallium/auxiliary/draw/draw_pt_varray.c
	src/gallium/auxiliary/draw/draw_pt_varray_tmp.h
	src/gallium/auxiliary/draw/draw_pt_vcache.c

+++ src/gallium/auxiliary/draw/draw_pipe_vbuf.c
@@ -145,7 +145,7 @@ emit_vertex( struct vbuf_stage *vbuf,
       vertex->vertex_id = vbuf->nr_vertices++;
    }
 
-   return vertex->vertex_id;
+   return (ushort)vertex->vertex_id;
 }
